requirements.txt
plotly
Pillow
pydantic